Move LogoutConfirmationDialog to feature:auth
