+++ settings.gradle
@@ -4,4 +4,4 @@ include ':data:repository'
 include ':data:db'
 include ':feature:album'
 include ':feature:artist'
-include ':feature:login'
+include ':feature:auth'
